+++ src/main/java/io/coodoo/workhorse/jobengine/control/JobEngineController.java
@@ -159,9 +159,18 @@ public class JobEngineController {
             Class<?> workerClass = Class.forName(job.getWorkerClassName());
             return (BaseJobWorker) CDI.current().select(workerClass).get();
 
+        } catch (ClassNotFoundException exception) {
+
+            logger.error("No JobWorker class found for {}", job);
+
+            job = jobEngineService.getJobById(job.getId());
+            job.setStatus(JobStatus.NO_WORKER);
+
+            throw exception;
+
         } catch (Exception exception) {
 
-            logger.error("Could not find JobWorker for job {}", job.getName(), exception);
+            logger.error("Could not instanciate JobWorker for job {}", job, exception);
 
             job = jobEngineService.getJobById(job.getId());
             job.setStatus(JobStatus.ERROR);
@@ -187,7 +196,8 @@ public class JobEngineController {
     }
 
     @TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
-    public synchronized JobExecution handleFailedExecution(Job job, Long jobExecutionId, Exception exception, Long duration, BaseJobWorker jobWorker) {
+    public synchronized JobExecution handleFailedExecution(Job job, Long jobExecutionId, Exception exception, Long duration, String jobExecutionLog,
+                    BaseJobWorker jobWorker) {
 
         JobExecution failedExecution = entityManager.find(JobExecution.class, jobExecutionId);
         JobExecution retryExecution = null;
@@ -221,7 +231,7 @@ public class JobEngineController {
         failedExecution.setStatus(JobExecutionStatus.FAILED);
         failedExecution.setEndedAt(JobEngineUtil.timestamp());
         failedExecution.setDuration(duration);
-        failedExecution.setLog(jobWorker.getJobExecutionLog());
+        failedExecution.setLog(jobExecutionLog);
         failedExecution.setFailMessage(exception.getMessage());
         failedExecution.setFailStacktrace(JobEngineUtil.stacktraceToString(exception));