+++ src/main/java/io/coodoo/workhorse/jobengine/control/event/JobErrorEvent.java
@@ -5,15 +5,24 @@ import io.coodoo.workhorse.jobengine.entity.Job;
 /**
  * @author coodoo GmbH (coodoo.io)
  */
-public class AllJobsDoneEvent {
+public class JobErrorEvent {
 
     private Job job;
 
-    public AllJobsDoneEvent(Job job) {
+    private Throwable throwable;
+
+    public JobErrorEvent(Job job, Throwable throwable) {
+        super();
         this.job = job;
+        this.throwable = throwable;
     }
 
     public Job getJob() {
         return job;
     }
+
+    public Throwable getThrowable() {
+        return throwable;
+    }
+
 }